rely on single base test versions from incubator openwhisk
