+++ tests/src/test/scala/system/basic/WskCliActionTests.scala
@@ -20,9 +20,26 @@ package system.basic
 import org.junit.runner.RunWith
 import org.scalatest.junit.JUnitRunner
 
+import common.TestCLIUtils
+import common.TestUtils.NOT_ALLOWED
 import common.Wsk
 
 @RunWith(classOf[JUnitRunner])
 class WskCliActionTests extends WskActionTests {
   override val wsk = new Wsk
+
+  it should "not be able to use --kind and --docker at the same time when running action create" in {
+    val file = TestCLIUtils.getTestActionFilename(s"echo.js")
+    val out = wsk.action.create(name = "kindAndDockerAction", artifact = Some(file), expectedExitCode = NOT_ALLOWED,
+        kind = Some("nodejs:6"), docker = Some("mydockerimagename"))
+    out.stderr should include("Cannot specify both --kind and --docker at the same time")
+  }
+
+  it should "not be able to use --kind and --docker at the same time when running action update" in {
+    val file = TestCLIUtils.getTestActionFilename(s"echo.js")
+    val out = wsk.action.create(name = "kindAndDockerAction", artifact = Some(file), expectedExitCode = NOT_ALLOWED,
+      kind = Some("nodejs:6"), docker = Some("mydockerimagename"), update = true)
+    out.stderr should include("Cannot specify both --kind and --docker at the same time")
+  }
+
 }
